MetaMo-Prototype/requirements.txt
python-dotenv
langgraph
langchain
langchain-core
langchain-community
langchain-openai
langchain-google-genai
requests
beautifulsoup4
